notes logic and mobile compatability
-finished notes app and endgame logic
-finished lost game logic
-added time to end game screen
-color adjustments

Perhaps unrelated to the amass of bug fixes and wrapping up this app, I'd like to introduce the concept of not coding like an asshole, which is simply defined as writing difficult to read about code.

More seriously, I think this project is a good example of why developers need to stick to consistent styling while trying their best to document as they go.